Add Rails 5.0; add back Rails 3.2 support
This adds both Rails versions to our test runs (3.2 was removed a short
time ago). I'm actually adding back Rails 3.2 as there are changes
needed for Rails 5 support that would cause a breaking release of
OkComputer (not compatible with Rails 3.2), but the amount of work
required to support 3.2-5.0 is minimal and I think that it isn't worth
forcing a major release on the gem.

And because it will make @anfleene squeal with glee.